build: allow locale directory to be passed in
A build request will fetch the most recent sources of the releng-tool
and translations repositories; however, a caller may already be working
on a translation repository already on the local system. Adjusting the
build request to allow a caller to pass in an already checked out source
of the translations repository.

Signed-off-by: James Knight <james.d.knight@live.com>