port over displayed version entries
The "releng-tool-doc-translations" repository no longer lists the
recommended order of versions to display in a final documentation build.
This repository will take over this list since (a) it was the only
repository who needs this information, and (b) it allows this repository
to customize which versions it wants to officially list in the version
list (aside from other/extra versions which may be built).

This commit drops some older versions and moves the "development"
version directly into the template (to force it on its own line).

Signed-off-by: James Knight <james.d.knight@live.com>