invoke releng-tool's setup call
The most recent version of releng-tool's documentation defines a
`setup` operation for Sphinx to process. The translation builder defines
its own setup operation and did not provide a means to support invoking
another setup call from the root documentation.

This commit tracks the root documentation's setup call (if one is
defined), and invokes it at the start of the setup call in the
translation builder's documentation configuration.

Signed-off-by: James Knight <james.d.knight@live.com>